lay out unit tests for the top-level platform module and for the addHelper module.
